Fix bug in expand_path_links
When consecutive pings are matched to links that are not directly
connected, a shortest path search is done to fill in the
intermediate links. While looping over the intermediate links
node ids from the outside scope were used to identify the key,
instead of the loop scope. Fix this and pull logic out to a
function.

This will likely not fix issues with missing links in the output.

Also fixes an unneeded (and costly) call to shortest path, use
the cache instead.