Enable skipping drivers
The query over the GPS table returns rows sorted by driver uuid
and timestamp so that pings may be grouped online by driver uuid
and are in the correct cronological order.

Update the gps query to optionally omit driver whose uuid
lexographically preceeds a given driver uuid. This enables the
map matching to be picked up (nearly) where it left off on a
previous run; or to start with a driver of interest

+ Remove unused kwarg from driver_segments
+ Add `--start-with-driver` and `--cache-path` to cli options
+ Enable the same level db to be used as the shortest path cache
  across runs
+ Format with black