Fix driver segment generator
`driver_segments` was not explicitly segmenting gps trajectories
on driver uuid. In the unlikely event that the end of one drivers
gps pings was within `min_duration` of the start of the next
drivers pings, then they would be returned as one segment.
This is now fixed.

`driver_segments` was also dropping rows that lay on the boundary
of two segments. This is now fixed.