Expose --accuracy-scale option
GPS pings come with a device-reported accuracy which is essentially
a confidence radius around the reported lat-lon. It is used to
compute the probability that a given ping should be matched to a
link based on its distance to it. Empirically it seems that the
GPS pings are less accurate than their reported accuracy would
suggest that they are so the --accuracy-scale option is added to
scale the accuracy values by the given factor. This has the effect
of making matches with distant links relatively more likely.

+ Improve accuracy repair process:
  - Convert negative accuracy to max trajectory accuracy
  - Set lower accuracy clip value as a constant
+ Consolidate pint unit registry in `common`
+ Change `LINK_SPEED` to avoid a name conflict with the GPS `speed`
  attribute
+ Improve SQL statment logging